refactor: StackView 활용으로 전환 (TableView -> StackView)
